Public interface for `publish/4` and `publish_sync/5`
Added a public interface for publishing messages on a connection
directly on the main `Tortoise`-module. This also allow the user to
reference the connection directly by `client_id` instead of a
`pipe`. Pipes will come back later.

Updated the getting started section in the Readme to use the
simplified interface.